Create transaction with hibernate

+++ src/main/java/com/bazarek/bazarek/dao/AddressDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressDao extends JpaRepository<Address,Integer> {
+public interface AddressDao extends JpaRepository<Address,Long> {
 }
